sonic_installer: fix read-only filesystem support for firmware update
It's possible that users will either store the SONIC image in a NTFS
formatted USB disk, or make it read-only by a USB lock.

Unfortunately the sonic_installer always tries to make the image executable
during the 'sonic_installer install' process, and then fail in these cases.

And thus, in this commit, we'll copy the image to /tmp upon chmod failure,
and then resume the following operations

Signed-off-by: Dante (Kuo-Jung) Su <dante.su@broadcom.com>