In the event of a kernel crash, we need to gather as much information as possible to understand and identify the root cause of the crash. Currently, the kernel does not provide much information, which make kernel crash investigation difficult and time consuming.
Fortunately, there is a way in the kernel to provide more information
in the case of a kernel crash. kdump is a feature of the Linux kernel
that creates crash dumps in the event of a kernel crash. This PR
will add kernel kdump support.

An extension to the CLI utilities config and show is provided to
 configure and manage kdump:
 - enable / disable kdump functionality
 - configure kdump
 - view kernel crash logs