Revert the change to kdump reboot (#916)
This is to revert the change to the reboot command in 7105400.

Platform reboot is not supported by the kdump kernel, as the platform drivers are not loaded. With the existing code, there are some error messages printed, but the device is still rebooted by CPU. The change here helps to avoid the error messages.