Use /var/tmp instead of /tmp for sonic_installer and generate_dump
sonic_installer and generate_dump utilities use currently /tmp directory
to store temporary files. Since the amount of data stored can be rather
large (especially in the case of the sonic_installer), we are switching to
use /var/tmp  instead.

We are doing these changes in anticipation that at some point /tmp will being
mounted over TMPFS

/tmp is typically mounted on TMPFS, which means it's fast and does not produce
wear on SSD or Hard-Drive. /var/tmp is mounted on the rootfs partition,
therefore not in the TMPFS partition.

We also noticed that both sonic_installer and generate_dump produce temporary
files which are not deleted once thse programes are done. We are then making
changes to peform a clean-up and remove the temporary files not needed anymore.

# Please enter the commit message for your changes. Lines starting
# with '#' will be ignored, and an empty message aborts the commit.
#
# Date:      Sat Jul 11 19:58:59 2020 -0400
#
# On branch var_tmp
# Changes to be committed:
#	modified:   scripts/generate_dump
#	modified:   sonic_installer/bootloader/aboot.py
#	modified:   sonic_installer/bootloader/grub.py
#	modified:   sonic_installer/bootloader/onie.py
#	modified:   sonic_installer/bootloader/uboot.py
#	modified:   sonic_installer/common.py
#	modified:   sonic_installer/main.py
#