[sfpshow] Gracefully handle improper 'specification_compliance' field (#1594)
Gracefully handle improper 'specification_compliance' field

The 'specification_compliance' field of transceiver info is expected to be a string representation of a dictionary. However, there is a chance, upon some kind of platform issue that a vendor's platform API returns something like 'N/A'. In this case, sfpshow would crash. Rather than crash, sfpshow should handle this gracefully and output 'N/A' instead.