[config] config reload needs to handle exceptions while stopping sonic.target
The below error is seen when a 'config reload' operation is performed
while the sonic.target is starting during system startup or a prior config
reload command.

Job for sonic.target canceled.

This error can be ignored if we can try to stop all the services associated
with sonic.target individually. Any processes which did not stop, will be
restarted as part of the "systemctl restart sonic.target" command.