[multi-asic][cli][chassis-db] Avoid connecting to chassis db
Currently, for all the cli commands, we connect to all databases
mentioned in the database_config.json. The database_config.json also
includes the databases from chassis redis server from supervisor card.
It is unneccessary to connect to databases from chassis redis server
when cli commands are executed form linecard. But we need to allow
connection to chassis databases when the cli commands are executed from
supervisor card.

The changes in this PR fixes this problem. The constructor of Db() class
which is instantiated for every CLI command execution is changed to skip
chassis databases from the list of collected databases if the card is not
supervisor card.

Signed-off-by: vedganes <vedavinayagam.ganesan@nokia.com>