[fast-reboot] revert the change of disabling counter polling before fast-reboot
The change was introduced with PR#1174 with the condition that flex counter
polling will be enabled in the boot up path unconditionally. However, a change
was made to enable_counters.py to only enable counters when the configuration
is missing (sonic-buildimage PR#7735). After the later change, the disabled
flex counters will not be enabled when fast-reboot into the same image.

Signed-off-by: Ying Xie <ying.xie@microsoft.com>