[show][config] cli refactor for muxcable with abstract class implementation from vendors (#1722)
Signed-off-by: vaibhav-dahiya vdahiya@microsoft.com

What I did
this PR adds support for cli refactor which is done for abstract class implementation of muxcable which is provided by vendors.
detailed design doc can be found Azure/SONiC#757

How I did it
Added the changes in muxcable.py in show and config .

How to verify it
add unit-tests as well as run on a Arista-7050cx3 switch, run unit tests.

no change in outputs, just the functionality has changed.
Signed-off-by: vaibhav-dahiya <vdahiya@microsoft.com>