[counters] Keep counters cache in a single directory
Place all counters cache in a single temporary repository. Then, when
swss starts in cold or fast mode we can clear this directory. Before
this change, every application is storing counters cache file in a
different location under different names, so it is hard to manage the
cache and know which files in /tmp need to be cleaned.

Added UserCache class which helps creating application specific per-user
cache directories.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>