[DPB] Enhance the code to shutdown interface before delete ports
Enhancement done in this commit:
Check the admin status on the ports to be deleted.
If the port admin status is up, do shutdown ports.
If the port admin status is down, do nothing.
Wait until the admin status of the ports to be removed on state DB has
changed to down.

Signed-off-by: chiourung_huang chiourung_huang@edge-core.com