[db_migrator] make LOG_LEVEL_DB migration more robust
It could be that LOG_LEVEL_DB includes some invalid data and/or a
KEY_SET that is not cleaned up due to an issue, for example we observed
_gearsyncd_KEY_SET set included in the LOG_LEVEL_DB and preserved in
warm reboot. However, this key is not of type hash which leads to an
exception and migration failure. The migration logic should be more
robust allowing users to upgrade even though some daemon has left overs
in the LOG_LEVEL_DB or invalid data is written.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>