[sonic-package-manager] Fix YANG validation failure on upgrade when feature has constraints in YANG model on FEATURE table
In case feature's YANG model has a constraint on FEATURE table:

```
must "current() = \"disabled\" or /feature:sonic-feature/feature:FEATURE/feature:FEATURE_LIST[feature:name=\"Y\"]/feature:state = \"enabled\"" {
    error-message "X can only be set when is Y enabled";
}
```

it will fail to upgrade due to transient disablement of the feature by
the upgrade process.

The upgrade process looks like this:
1. Disable feature via FEATURE table and wait for container to go down.
2. Remove all configuration files generated for this feature.
3. Create new configuration files for new docker image.
4. Merge new init_cfg from the new docker image and perform YANG
   validation.
5. Enable feature via FEATURE table.
6. Remove old docker image.

Since there is a transient enable/disable of the feature and YANG
validation in the middle having the above constrains in YANG model would
lead to an upgrade failure.

I attempt to fix this by letting the ServiceCreator know we undergo a
transient disablement of the feature.

Signed-off-by: Stepan Blyschak <stepanb@nvidia.com>