Fix test execution on Bookworm (#3041)
* Fix test execution on Bookworm

Running tests with `python3 -m pytest` instead of `python setup.py test`
causes some breakages on Bookworm. Some of this has to do with what
module searc paths get added, while others appear to be related to
changes in the `mock` module between Python 3.9 and Python 3.11.

Fix this and make sure it works on both Bullseye and Bookworm.

* Fix mclag_test.py having different results for some test cases

The test_mclag_add_mclag_member_to_nonexisting_domain test case passes
when run on Bullseye, but fails on Bookworm. This is because some test
cases modify the value of `mclag.ADHOC_VALIDATION`, but this value may
persist for other test cases as well, and if test cases happen to run in
a different order, then it may unexpectedly fail.

For now, fix `test_mclag_add_mclag_member_to_nonexisting_domain` by
setting the value there.

* Add check_output parameter to the setup function due to the patch

Since there is a patched function specified as an attribute, newer
versions of mock expect that the object can be passed in as a parameter
to the function. However, the `setup` functions didn't accept it as a
parameter.

Modify the `setup` functions to accept a parameter for this object.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>

---------

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>