Fix test execution on Bookworm
Running tests with `python3 -m pytest` instead of `python setup.py test`
causes some breakages on Bookworm. Some of this has to do with what
module searc paths get added, while others appear to be related to
changes in the `mock` module between Python 3.9 and Python 3.11.

Fix this and make sure it works on both Bullseye and Bookworm.