Modify teamd retry count script to base BGP status on default BGP status
For each BGP status, if the `admin_status` field is not present, then
whether the BGP session is admin up or admin down depends on the default
BGP status (in the `default_bgp_status` field coming from
`init_cfg.json`), which is specified during image build. If the default
BGP status is up, then `admin_status` will be created only when the BGP
session is brought down; similarly, if the default BGP status is down,
then `admin_status` will be created when the BGP session is brought up.

Because of that, modify the script to use the default BGP status as the
initial value.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>