Optimize lag_keepalive by crafting the LACPDU packet ourselves
Instead of waiting for a LACPDU packet to be sent and capturing that
(which involves waiting roughly 30 seconds), get the necessary
information from teamd and craft it ourselves. This means that the
60-second wait in making sure that a LACPDU packet is captured and the
keepalive script is ready can be largely eliminated (this has been
reduced to 10 seconds to make sure the script has a chance to craft the
packets and send some LACPDUs).

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>