Don't exit immediately if running a command under alias mode
There's a difference in behavior when an external command is run under
the default mode vs when it is run under the alias mode. In the default
mode, execution control returns to the caller unless the command had a
non-zero exit code. In the alias mode, regardless of exit code, the
Python script exits. This may result in some tasks not completing.

Fix this by not unconditionally exiting if running a command in the
alias mode. Note that there are other differences still present, but
this fixes at least this one.

Signed-off-by: Saikrishna Arcot <sarcot@microsoft.com>