sonic-installer: enhance next image detection for Aboot
The Aboot bootloader relies of the SWI= keyword argument in the
boot-config file to know which image to boot.
This value is also used by sonic-installer to figure to extract the next
image that will be executed.
The current code has an issue as it only expects the next image to match
the installation path of a SONiC image but not anything else.

This means that `SWI=flash:sonic-aboot-broadcom.swi` is not valid and
can therefore be a problem when trying to install a new image via cold
reboot.

Additionally a missing or empty boot-config would generate a python
backtrace instead of gracefully recovering from this state.